IMAPTalk: fix authentication to non-SASL-IR servers
4.03 introduced support for authenticating on behalf of a different user
using AUTHENTICATE when the remote server supports AUTH=PLAIN and
SASL-IR.

Unfortunately, the test to decide whether to do LOGIN or AUTHENTICATE
was broken (AUTH=LOGIN refers to the SASL LOGIN mechanism, not the IMAP
LOGIN command, which _always_ exists). As a result, we were always using
AUTHENTICATE PLAIN, even when no AuthzUser was provided.

This is fine most of the time, except in the rare cases where a server
supports AUTH=PLAIN but doesn't support SASL-IR. In those cases, we
would crash.

This fix makes it always use LOGIN except with AuthzUser is provided, in
which case it uses AUTHENTICATE PLAIN instead.