Mail::IMAPTalk: add ->unselect_or_close
seznam.cz is a big email provider in Czechia, but has (to quote an
expert) a terrible IMAP implementation.  For one thing, it doesn't
implement UNSELECT, which we use without checking for it.

We want the migrator to work, so we need to cope with this.  This method
adds ->unselect_or_close which will use UNSELECT if advertised, and
CLOSE otherwise.  We connect to remote IMAP in basically two places:

1.  the migrator, which does not affect mail in situ, and always(?) uses
    EXAMINE
2.  MailFetch-via-IMAP, which *does* remove mail from the remote by
    marking it \Deleted and later making issuing a CLOSE

So, if we switch more or less everything that uses UNSELECT to use this
method, the only affected case should be remote IMAP servers involved in
a MailFetch, where the target mailbox has no new mail that would be
deleted but has *existing* old mail already flagged \Deleted.  In this
case, that mail might be expunged -- but even then probably not, since
we should not be issuing a CLOSE *or* UNSELECT.  Still, Mail::IMAPTalk
issues unselects fairly freely, so this might happen.